Move LogLevel into a new class file
